feat : ENTITY 수정본
