+++ src/main/java/com/example/outsourcing/controller/LogController.java
@@ -1,4 +1,4 @@
-package controller;
+package com.example.outsourcing.controller;
 
 import org.springframework.stereotype.Controller;
